pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecg-pipeline"
version = "0.1.0"
description = "Memory-lean ECG denoising pipeline with IndexMap projections"
authors = [{name="You", email="you@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "matplotlib",
  "scipy",
  "pyyaml",
  # neurokit2 is optional at runtime if you choose method: 'neurokit'
  "neurokit2; python_version >= '3.9'",
]

[project.scripts]
ecg-pipeline = "cli:main"

[tool.setuptools]
packages = ["ecg_pipeline"]